perf: shrink matcher bitmasks to 32 bits
Needle masks use only bits 0 through 25, while candidate masks use a
lossy 32-bucket mapping. Store both in `uint32_t` instead of `long` to
shrink each `haystack_t` from 32 bytes to 24 bytes.

A candidate can legitimately set all 32 bucket bits, so `UINT32_MAX`
cannot serve as its unset sentinel. Reserve bit 31 as a computed flag
instead. Needle masks never test that bit, so it does not affect subset
filtering, and non-letter collisions remain harmless false positives.
Keep a separate `UINT32_MAX` sentinel for needle masks, whose valid
values use only the lower 26 bits.

This preserves the explicitly wrapping candidate mapping and the
scorer's established control flow while reducing the 400k-candidate
haystack slab by approximately 3.2 MB.

Validation completed with 168 passing tests and 1 skip in release,
debug, UBSan, ASan, and GCC 14 builds. The full benchmark supplied by
the user was:

```
                    best    avg      sd     +/-      p     (best)    (avg)      (sd)     +/-      p
     pathological 0.63949 0.67828 0.03816 [-1.3%]        (0.63948) (0.67768) (0.03754) [-1.2%]
        command-t 0.10950 0.12826 0.02141 [+4.8%]        (0.10950) (0.12790) (0.02114) [+4.7%]
chromium (subset) 0.61022 0.64994 0.01437 [-3.1%]  0.005 (0.12027) (0.12830) (0.00331) [-3.3%] 0.0005
 chromium (whole) 0.33049 0.33420 0.00391 [-4.6%] 0.0005 (0.04279) (0.04768) (0.00270) [-2.7%]
       big (400k) 0.55969 0.56559 0.00476 [-4.9%] 0.0005 (0.07181) (0.07660) (0.00267) [-6.2%] 0.0005
            total 2.25943 2.35627 0.04826 [-2.8%]  0.005 (0.99548) (1.05817) (0.04440) [-1.3%]
```

Agent prompts used in preparing this commit:

> is there any point in looking at reducing the bitmask size in the
> struct now?

> try it

> it's good. commit this. include these benchmark results in the message:

+++ lua/wincent/commandt/lib/commandt.h
@@ -19,7 +19,7 @@
  */
 typedef struct {
     str_t *candidate;
-    long bitmask;
+    uint32_t bitmask;
     float score;
 
     /**
@@ -179,7 +179,7 @@ typedef struct {
      */
     const char *needle;
     size_t needle_length;
-    long needle_bitmask;
+    uint32_t needle_bitmask;
 
     const char *last_needle;
     size_t last_needle_length;
